backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
web3>=6.15.0
requests>=2.32.0
numpy>=2.0.0
slowapi>=0.1.9
python-multipart>=0.0.6

